chore(index): Create new html file, fill-out and create link to return
Iteration 2: recipe page

Created a new recipe file for the pepper steak and added a Home link that brings back the user to the main page.
Made some changes in the steps on the README.md file.
Proceeded with the HTML boilerplate for the new HTML file, and connected both index and recipe page, together with The Odin Project steps.