chore(peppersteak): Add Image, description, ingredients, and Steps Iteration 3: recipe page content
In the Pepper Steak page I added an beautiful image of a cooked and served pepper steak dish. All together with the image I added an alt description for the image, so that readers can read it.
Next three things was the headline description with the paragraph separated each sentence on a new line. Added detailed ingredients in an unordered list and the steps to prepare in an ordered one.